Get Citymapper's mobility index

As of mid-March Citymapper has launched a City Mobility index to observe which cities are following lockdown measures during COVID-19 pandemic. Data is calculated comparing the number of trips planned on the app to a typical usage period. Here for a detailed description.

get_mob_idx(
  start_date = "2020-01-01",
  end_date = Sys.Date(),
  city = NA,
  cache = TRUE,
  weekly = FALSE,
  verbose = FALSE
)

Arguments

start_date
First observed day with format YYYY-MM-DD

end_date
Last observed day with format YYYY-MM-DD

city
(optional) A vector containing specific cities to filter for (e.g "milan", lower-case). This obviously follow Citymapper's coverage.

cache
(bool) cache results to filesystem

weekly
Get weekly data (default is daily)

verbose
Show download progress bar

Value

A data frame containing the following columns:

big_change: (bool) Whether there was a consisten change with previous day.

cange_from_previous: Difference with previous day.

date: Observation date

type: Weekly or daily observation.

region_id: Region slug.

value: Index value.

city_name: Name of oberved city